sửa tạo bài viết theo danh mục đấu giá và các loại khác tách file views sửa thông báo và thêm action đã xem thông báo sửa model thêm upload ảnh lên cloud
