+++ BackEnd/CharitySocialNetwork/AppCharitySocialNetwork/Templates/index.html
@@ -32,7 +32,7 @@
                     <a href="https://swagger.io/tools/swagger-ui/" class="">
                         <div class="card ">
                             <span class="badge badge-danger ">V1</span>
-                            <img class="card-img-top p-4" src="/static/images/index/swagger_logo.svg" alt="">
+                            <img class="card-img-top p-4" src="/static/assets/images/index/swagger_logo.svg" alt="">
                             <div class="card-body">
                                 <h5 class="card-title text-center" style="color:#ff6600d6">Swagger API Debug</h5>
                             </div>
@@ -43,7 +43,7 @@
                     <a href="https://redoc.ly/docs" class="">
                         <div class="card ">
                             <span class="badge badge-danger ">V1</span>
-                            <img class="card-img-top p-4" src="/static/images/index/redoc-logo.png" alt="">
+                            <img class="card-img-top p-4" src="/static/assets/images/index/redoc-logo.png" alt="">
                             <div class="card-body">
                                 <h5 class="card-title text-center" style="color:#ff6600d6">Reference Documentation</h5>
                             </div>
@@ -55,7 +55,7 @@
                         <div class="card ">
                             <span class="badge badge-danger ">V1</span>
                             <img class="card-img-top p-4" style="background: #28a745"
-                                 src="/static/images/index/logo-django.svg" alt="">
+                                 src="/static/assets/images/index/logo-django.svg" alt="">
                             <div class="card-body">
                                 <h5 class="card-title text-center" style="color:#ff6600d6">Django</h5>
                             </div>
@@ -66,7 +66,7 @@
                     <a href="https://www.django-rest-framework.org/api-guide/requests/" class="">
                         <div class="card ">
                             <span class="badge badge-danger ">V1</span>
-                            <img class="card-img-top p-4" src="/static/images/index/logo_resframework.png" alt="">
+                            <img class="card-img-top p-4" src="/static/assets/images/index/logo_resframework.png" alt="">
                             <div class="card-body">
                                 <h5 class="card-title text-center" style="color:#ff6600d6">Django Res Framework</h5>
                             </div>
@@ -77,7 +77,7 @@
                     <a href="https://graphql.org/code/#python" class="">
                         <div class="card ">
                             <span class="badge badge-danger ">V1</span>
-                            <img class="card-img-top p-4" src="/static/images/index/logo graphql.svg" alt="">
+                            <img class="card-img-top p-4" src="/static/assets/images/index/logo graphql.svg" alt="">
                             <div class="card-body">
                                 <h5 class="card-title text-center" style="color:#ff6600d6">Django Graphql</h5>
                             </div>
@@ -129,7 +129,7 @@
                     <a href="/api" class="">
                         <div class="card ">
                             <span class="badge badge-danger ">V1</span>
-                            <img class="card-img-top p-4" src="/static/images/index/logo_resframework.png" alt="">
+                            <img class="card-img-top p-4" src="/static/assets/images/index/logo_resframework.png" alt="">
                             <div class="card-body">
                                 <h5 class="card-title text-center" style="color:#ff6600d6">Charity Social Network
                                     API</h5>
@@ -141,7 +141,7 @@
                     <a href="/swagger" class="">
                         <div class="card ">
                             <span class="badge badge-danger ">V1</span>
-                            <img class="card-img-top p-4" src="/static/images/index/swagger_logo.svg" alt="">
+                            <img class="card-img-top p-4" src="/static/assets/images/index/swagger_logo.svg" alt="">
                             <div class="card-body">
                                 <h5 class="card-title text-center" style="color:#ff6600d6">Charity Social Network
                                     Swagger</h5>
@@ -153,7 +153,7 @@
                     <a href="/redoc" class="">
                         <div class="card ">
                             <span class="badge badge-danger ">V1</span>
-                            <img class="card-img-top p-4" src="/static/images/index/redoc-logo.png" alt="">
+                            <img class="card-img-top p-4" src="/static/assets/images/index/redoc-logo.png" alt="">
                             <div class="card-body">
                                 <h5 class="card-title text-center" style="color:#ff6600d6">Charity Social Network
                                     Reference Documentation</h5>
@@ -165,7 +165,7 @@
                     <a href="/admin" class="">
                         <div class="card ">
                             <span class="badge badge-danger ">V1</span>
-                            <img class="card-img-top p-4" src="/static/images/index/" alt="">
+                            <img class="card-img-top p-4" src="/static/assets/images/index/" alt="">
                             <div class="card-body">
                                 <h5 class="card-title text-center" style="color:#ff6600d6">Charity Social Network
                                     Admin</h5>
@@ -180,6 +180,7 @@
 
 
     </div>
+    <img src="https://www.dropbox.com/s/57k3qm85byoewb1/226192460_417920722869038_2260780749190981076_n.png?dl=0#">
 
 </div>
