+++ BackEnd/CharitySocialNetwork/AppCharitySocialNetwork/view/__init__.py
@@ -1,0 +1,6 @@
+from .historyauctionview import HistoryAuctionViewSet
+from .postview import PostViewSet
+from .userview import UserView
+from .reportview import ReportViewSet
+from .emotionview import EmotionTypeViewSet
+from .viewcomment import CommentViewSet
